Moved classes to better packages, refactoring, tweaked the test
