+++ subprojects/extTest/src/test/resources/mockito-extensions/org.mockito.plugins.InstantiatorProvider2
@@ -1,1 +1,1 @@
-org.mockitousage.plugins.switcher.MyInstantiatorProvider2
+org.mockitousage.plugins.instantiator.MyInstantiatorProvider2
